Partial change-set to make MoSync compile on AMD64 debian linux.
Some changes are hacks, one definitely needs fixing: tools/package/windowsPhone.cpp

+++ intlibs/stabs/stabs_symbol.h
@@ -32,7 +32,7 @@ struct Symbol {
 	bool global;
 	int fileScope;
 	int lineNumber;	//on which this symbol is defined
-	int address;
+	size_t address;
 };
 
 #endif	//STABS_SYMBOL_H
